Move the core provider types to pkg/cloud/types
This is a first step to enable moving providers into their own
packages - it avoids dependency cycles such as:
pkg/cloud.GetProvider -> pkg/cloud/azure.Azure ->
pkg/cloud.CustomPricing
